upgrade: image node and ImageView to tiptap2
Quite a bit of the syntax changed.
We now need a wrapping   `<node-view-wrapper>` element.
Pretty good docs at https://tiptap.dev/guide/node-views/vue#render-a-vue-component

We also need to handle the async action.
It will run the action on it's own.
So in `clickIcon()` we need to test if the action returned anything.