2nd-try: fix linked files not working in richworkspace editor.
Introduce a global property

OCA.Text.RichWorkspaceFilePath

which is used in helpers/links.js if the Viewer-property OCA.Viewer.file
is not available.

This fixes the problem that opening linked files from the enriched
workspace does not work, as OCA.Viewer.file is empty in this case. So
somehow the base MD-file has to be to communicated to domHref() in
links.js.

Signed-off-by: Claus-Justus Heine <himself@claus-justus-heine.de>